Break up the module structure a bit, add tests
