+++ src/lib.rs
@@ -1,3 +1,6 @@
+pub mod matchers;
+pub mod prelude;
+
 pub struct ExpectationTarget<A> {
     target: A
 }
@@ -10,38 +13,10 @@ impl <A> ExpectationTarget<A> {
     }
 }
 
-
-pub struct EqualityMatcher<E> {
-    expected: E
-}
-
-
-impl <E, A> Matcher<A> for EqualityMatcher<E> where E: PartialEq<A> {
-    fn matches(self, target: &A) -> bool {
-        &self.expected == target
-    }
-}
-
-pub struct BeAnEmptyVectorMatcher;
-
-impl <A> Matcher<Vec<A>> for BeAnEmptyVectorMatcher {
-    fn matches(self, target: &Vec<A>) -> bool {
-        target.len() == 0
-    }
-}
-
 pub trait Matcher<A> {
     fn matches(self, target: &A) -> bool;
 }
 
 pub fn expect<A>(target: A) -> ExpectationTarget<A> {
     ExpectationTarget { target: target }
-}
-
-pub fn eq<E>(expected: E) -> EqualityMatcher<E> {
-    EqualityMatcher { expected: expected }
-}
-
-pub fn be_an_empty_vector() -> BeAnEmptyVectorMatcher {
-    BeAnEmptyVectorMatcher
 }
