+++ src/prelude.rs
@@ -1,0 +1,2 @@
+pub use matchers::prelude::*;
+pub use super::expect;
